Create README.md
This code is for accelerometer and pulse rate sensor. when a person is sleeping his/her heart beat becomes slow. Therefore it switches on the vibrator inthsi case. also if a person is not moving his hands while driving..that means he/she is sleeping. accelerometer will check this and phenomenon and vibrator is switched on inr response to it. Basically this code is to create a smart device which will prevent road accidents.